Add Watchman service component
This component is designed to create a list of EXPECTED endpoints given a config file
and check if those endpoints exist/are healthy. Rather than pinging Ambassador directly,
we want to have a list of endpoints the user was expecting, if something happens between
actually deploying the model we can report the endpoint is not up.

Can later be extended to report additional information as well per endpoint / model / project level