Add RandomDataProvider
Also changes RandomDataset to be based on TimeSeriesDataset with
RandomDataProvider as the data source.  As a consequence RandomDataset
now requires a start and end-time, and a list of tags.

Also fix logger in providers.py. It was imported from
`gordo_components.dataset.datasets`, now it is
`logger = logging.getLogger(__name__)` as it should be