Generalize kwargs in models
Add kwargs to Keras' fit_generator in KerasLSTMAutoEncoder fit method,
so arguments supplied to the fit_generator will be updated as per
the kwargs supplied when instantiating the class.

This excludes the keyword argument "shuffle" which is hard coded to
false, given that we are dealing with time series data.