Add support for prefix less tags
Doing this by having an optional "asset" flag in the config. If nothing else changes,
this asset flag is set on each tag in the list.
In gordo-components, a tag changes from being a string, to being a namedtuple.
If asset is not specified then it is just None, and not considered when resolving paths.
In the serialized output (metadata) the namedtuple is represented as a dict.

Add testing of NcsReader via mocking

Add normalizing function of tag lists, use in TimeSeriesDataset init