Watchman: Listen to updated services
Listens to k8s for updated events regarding services, and reacts
to them instead of scanning the list of targets. Listens in a background-thread
and schedules regular jobs for updating the metadata when it sees a new service.
When the metadata is successfully updated it will reschedule the task to happen
regularly, but rarely just in case there is any change to meta data without
a change in the service-description.