Replace generate_window with keras functionality
Replaces it with keras TimeseriesGenerator, and created a function
create_keras_timeseriesgenerator which also allows for having
y shifted arbitrary relative to X. By default TimeseriesGenerator only
allows for a shift of 1.

Sets 10000 as default batch_size on predict. Batch_size on predict is not the
same as the on for training, and is purely a performance trade-of
(memory vs speed). Thus it is not customizable from yaml. Did some rough
benchmarks with different loookback-windows (144 / 72 / 36) and
n_features (3,6) on transforming 100000 datapoints,
and using a batch-size around 10000 was the fastest in general.

Made it such that all calls to fit calls super().fit, which generates a
completely fresh model. This ensures that we are compatible with sklearn,
which expects this behavior.

The new method might use a bit more memory. The reason is that we have to
insert upto several rows at the start and end of the dataset for
TimeseriesGenerator to generate the lookahead, and that probably causes numpy
to copy the whole matrix.

Currently there is a small change in that the first sample will be used twice
in fit, once to initialize the model and then again in fit_generator.