Support predicting n number of tags
Update serializer to recognize parameters as possible path.to.Model
which will be loaded and passed back to the class specifying that
path. Such that Model(estimator="sklearn.fancy.FancyModel") will
be loaded as Model(estimator=FancyModel()).

This then allows using things like sklearn.multioutput.MultiOutputRegressor
which needs an 'estimator' as a parameter in the config file specification.