Add multithreaded download of tags from NCS
Also exposed the 'threads' parameter in the `DataLakeProvider`
with the contract that a value of `None` gives the underlying
dataprovider the freedom to choose the number of threads it sees fit.
`None` is the default value of `threads` in DataLakeProvider, but
both the model builder and the client sets it to 1 unless specified
by the user, so as to retain the old behavior.