TimeSeriesDataset: custom aggregation method(s)
Now the TimeSeriesDataset supports an optional argument `aggregation_methods`
with default value `mean`. It can be either a single aggregation method or
a list of aggregation methods. Any aggregation method supported by pandas
is supported. If multiple aggregation methods are provided then the resulting
dataframe contains a multi-level column index with the tag-name as the top
level, and the aggregation method as the second level. If a single aggregation
method is provided then only the first level (tag-name) is used.