Fix logging-issue stemming from absl
Tensorflow imports absl, and it breaks the standard logging module in
python by setting a handler on the root-level logger. The result is
that the log-level does not get set, and the format is all wrong.
This patch introduces an ugly fix I found on the internet to handle it.
Hopefully we can delete it in the future.

References:
https://github.com/tensorflow/tensorflow/issues/29842
https://github.com/tensorflow/tensorflow/issues/26691
https://github.com/abseil/abseil-py/issues/99