Limit the layer between raw configs to object initializations
Much of the raw config preprocessing occurs in the CLI, outside
of the context of builder.build_model. Things like renaming "train_start_date"
inside of this CLI makes it quite confusing and opaque to the user of how
the raw configs are translated into the object used to get the data
and build the models.

The removes or at least condenses this logic into the objects themselves,
rather than doing the preprocessing between the time the yaml is submitted
and when the objects are initialized from those configs.