Add ability to scale y when scoring models (#599)
The evaluation-config dictionary passed to the model-builder now takes
a new parameter, scoring_scaler, which can be a scaler, or None. If it is set
then it is fitted to `y` in the model-building phase, and applied to the data passed
to the metrics before the metrics is calculated. This is useful if the metrics are over
different series on vastly different magnitudes, and the metrics is sensitive to this. 
Default value is `sklearn.preprocessing.RobustScaler`.