Workflow: Clean up quoting data to builder
This standardizes on using the jinja2 json filter and yaml literal
quote to send data to the model-builder for all data. Earlier we relied
on python string-formatting the values in a sane way, which did not work
if the strings contained a single quote, since then python would switch to
using double quotes where it usually uses single quotes.

This closes #682